+++ src/main/java/com/example/jejuiiin/controller/handler/GlobalExceptionHandler.java
@@ -5,7 +5,7 @@ import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
 
-import static com.example.jejuiiin.exception.ExceptionMessage.INVALID_PAGE_NUMBER_MSG;
+import static com.example.jejuiiin.controller.exception.ExceptionMessage.INVALID_PAGE_NUMBER_MSG;
 import static org.springframework.http.HttpStatus.BAD_REQUEST;
 
 @RestControllerAdvice
