+++ src/main/java/com/example/jejuiiin/security/jwt/exception/CustomSecurityException.java
@@ -1,6 +1,6 @@
 package com.example.jejuiiin.security.jwt.exception;
 
-import com.example.jejuiiin.exception.ExceptionMessage;
+import com.example.jejuiiin.controller.exception.ExceptionMessage;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
